Removed baseexperiment.h and basedatatype.h and replace themn with utilityenum.h ; Also, move output file function to utils.h
