+++ src/utilityenum.h
@@ -1,0 +1,27 @@
+#ifndef BASEEXPERIMENT_H
+#define BASEEXPERIMENT_H
+
+#include <ifloat.h>
+#include <QString>
+#include <vector>
+#include <fstream>
+#include <map>
+using std::vector;
+
+enum Op{ADD, SUB, MUL, DIV, MATMUL};
+
+enum DataType{NUMERIC, ARRAY, MATRIX};
+
+enum Algo {LINEAR, SPLIT_MERGE, SORT, SORT_APPEND, GROUND_TRUTH};
+
+struct Result {
+    iFloat value;
+    Algo algoUsed;
+    Result(iFloat v, Algo algo) {value = v; algoUsed = algo;}
+
+    Result() {value = 0; algoUsed = GROUND_TRUTH;}
+};
+
+
+
+#endif // BASEEXPERIMENT_H
